use host IP to run on android physical device
